Add user auth form and upload form
Pair Programmed: Kat and Matt
Adds auth form in index.html and adds events, api, ui javascript files
to allow users to sign up, sign in successfully. This is a skeleton so
some ui code is commented out for now.
Add upload form for users to upload a file. And adds api, events, ui
javascript files to connect to the api.