Add Edit File Functionality
Add event listener for edit button, which allows a user to input a new
file name, as well as tags for a file. When a user confirms the changes,
a patch request is sent to the API, which modifies the data in MongoDB.
When this action is successful, the table automatically refreshes with
the user's changes.

Full team programming