Add Working, Clickable Hashtags
When editing the tags for a file, clicking confirm separates the tags by
the spaces in between them, creating an array of strings, and updating
the file tags on the API.

In Handlebars, adds logic to parse the array of strings for hashtags,
rendering a clickable link as well as the # symbol in front of the tag.

Clicking one of the hashtags sends a GET request to /uploads, passing in
a query string of the hashtag that was clicked. The API then returns
only the files that contain that particular hashtag.