Add Styling Fixes, Fix Table Refresh Btn
Changed the sizing of the file input box to ensure that it stays on one
line for most browser displays.

Refactored the on page load code in index.js.

Fixed the refresh table button so that the event listener is activated
once the table is generated, and not before.

Pair Programmed with Kat